네트워크 SVG 이미지 화면에 띄우기
