+++ Neil Chaeseong/MultipartTest/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Wed Mar 29 10:03:37 KST 2023
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.5-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
