--- Neil Chaeseong/OauthPractice/app/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">OauthPractice</string>
-</resources>
