--- Neil Chaeseong/OauthPractice/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun Mar 19 12:07:20 KST 2023
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.5-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
